[cloud] use 1 deployment for multiple simulations (#67)
* [cloud] use deployment for multiple simulations

* fix compile

* print mvn/java version

* fix mvn command, add cleanup class

* print logs

* downgrade scala maaven plugin

* scala:run

* try very old plugin version

* try running compile

* check maven version

* simplify pipeline

* check mvn version in main container

* run container as root

* run via exec:java

* add thread split

* fix parallel call

* adjut script

* adjust logging

* MAX_TEST_RUNS=10

* [ci] set 90 min timeout for cloud pipeline

* avoid resetting password

* fix logs

* remove calls for host and version

* do not delete deployment after 1 sim

* fix compilation

* fix parameter message

* update config

* fail script faster

+++ src/test/scala/org/kibanaLoadTest/simulation/BaseSimulation.scala
@@ -2,6 +2,7 @@ package org.kibanaLoadTest.simulation
 
 import io.gatling.core.Predef._
 import io.gatling.core.structure.ChainBuilder
+import io.gatling.http.protocol.HttpProtocolBuilder
 import org.kibanaLoadTest.KibanaConfiguration
 import org.kibanaLoadTest.helpers.{
   CloudHttpClient,
@@ -37,7 +38,10 @@ class BaseSimulation extends Simulation {
       )
   }
 
-  SimulationHelper.randomWait
+  // -DdeploymentId=67d0195ac345578bcab6e561ff5, optional to use existing deployment
+  val deploymentId: Option[String] = Option(
+    System.getProperty("deploymentId")
+  )
   // -DdeploymentConfig=path/to/config, default one deploys basic instance on GCP
   val CLOUD_DEPLOY_CONFIG: String =
     System.getProperty("deploymentConfig", "config/deploy/default.conf")
@@ -48,7 +52,13 @@ class BaseSimulation extends Simulation {
   // -DenvConfig=path/to/config, default is a local instance
   val envConfig: String = System.getProperty("env", "config/local.conf")
   // appConfig is used to run load tests
-  val appConfig: KibanaConfiguration = if (cloudDeployVersion.isDefined) {
+
+  val appConfig: KibanaConfiguration = if (deploymentId.isDefined) {
+    logger.info(s"Using existing deployment: ${deploymentId.get}")
+    SimulationHelper
+      .useExistingDeployment(deploymentId.get)
+      .syncWithInstance()
+  } else if (cloudDeployVersion.isDefined) {
     // create new deployment on Cloud
     logger.info(s"Reading deployment configuration: $CLOUD_DEPLOY_CONFIG")
     SimulationHelper
@@ -63,9 +73,9 @@ class BaseSimulation extends Simulation {
       .syncWithInstance()
 
   val httpHelper = new HttpHelper(appConfig)
-  var httpProtocol = httpHelper.getProtocol
-  var defaultHeaders = httpHelper.getDefaultHeaders
-  var defaultTextHeaders = httpHelper.defaultTextHeaders
+  var httpProtocol: HttpProtocolBuilder = httpHelper.getProtocol
+  var defaultHeaders: Map[String, String] = httpHelper.getDefaultHeaders
+  var defaultTextHeaders: Map[String, String] = httpHelper.defaultTextHeaders
 
   if (appConfig.isSecurityEnabled) {
     defaultHeaders += ("Cookie" -> "${Cookie}")
@@ -91,7 +101,9 @@ class BaseSimulation extends Simulation {
   }
 
   after {
-    if (appConfig.deploymentId.isDefined) {
+    if (
+      appConfig.deploymentId.isDefined && appConfig.deleteDeploymentOnFinish
+    ) {
       // delete deployment
       new CloudHttpClient().deleteDeployment(appConfig.deploymentId.get)
     } else {
